Fix removal of negation when no prefix
Previously, when the user tried to remove negation using backspace, in a
control having no prefix or format, the onChange handler was not being
called.

This was caused by a related fix, allowing the removal of negation when
there was a prefix. The value was being updated manually in onKeyDown,
which would leave nothing to change for the browser handling of the key
event, and so no change event was being emitted.

The fix was to only apply the previous code when a prefix was defined.